DRY-ify expert_demos and rename to train_rl
