--- docs/modules/imitation.scripts.config.rst
@@ -1,41 +1,0 @@
-imitation.scripts.config package
-================================
-
-Submodules
-----------
-
-imitation.scripts.config.analyze module
-----------------------------------------
-
-.. automodule:: imitation.scripts.config.analyze
-
-imitation.scripts.config.common module
---------------------------------------
-
-.. automodule:: imitation.scripts.config.common
-
-imitation.scripts.config.eval\_policy module
---------------------------------------------
-
-.. automodule:: imitation.scripts.config.eval_policy
-
-imitation.scripts.config.expert\_demos module
----------------------------------------------
-
-.. automodule:: imitation.scripts.config.expert_demos
-
-imitation.scripts.config.parallel module
-----------------------------------------
-
-.. automodule:: imitation.scripts.config.parallel
-
-imitation.scripts.config.train\_adversarial module
---------------------------------------------------
-
-.. automodule:: imitation.scripts.config.train_adversarial
-
-
-Module contents
----------------
-
-.. automodule:: imitation.scripts.config
