+++ docs/modules/imitation.scripts.rst
@@ -1,13 +1,6 @@
 imitation.scripts package
 =========================
 
-Subpackages
------------
-
-.. toctree::
-
-   imitation.scripts.config
-
 Submodules
 ----------
 
@@ -21,11 +14,6 @@ imitation.scripts.eval\_policy module
 
 .. automodule:: imitation.scripts.eval_policy
 
-imitation.scripts.expert\_demos module
---------------------------------------
-
-.. automodule:: imitation.scripts.expert_demos
-
 imitation.scripts.parallel module
 -------------------------------------------
 
@@ -36,6 +24,25 @@ imitation.scripts.train\_adversarial module
 
 .. automodule:: imitation.scripts.train_adversarial
 
+imitation.scripts.train\_bc module
+----------------------------------
+
+.. automodule:: imitation.scripts.train_bc
+
+imitation.scripts.train\_dagger module
+--------------------------------------
+
+.. automodule:: imitation.scripts.train_dagger
+
+imitation.scripts.train\_preference_comparisons module
+------------------------------------------------------
+
+.. automodule:: imitation.scripts.train_preference_comparisons
+
+imitation.scripts.train\_rl module
+----------------------------------
+
+.. automodule:: imitation.scripts.train_rl
 
 Module contents
 ---------------
